vexpress, sunxi: move common MMU TTB code to core/arch/arm/mm
