+++ core/arch/arm/plat-sunxi/conf.mk
@@ -16,6 +16,7 @@ CFG_SECURE_TIME_SOURCE_CNTPCT := y
 CFG_WITH_SEC_MON := y
 CFG_WITH_STACK_CANARIES := y
 CFG_SUNXI_UART ?= y
+CFG_MMU_V7_TTB ?= y
 
 include mk/config.mk
